Refactor and comment gpt-tfjs implementation
